+++ rulesets/std.py
@@ -6,6 +6,6 @@ STD_RULES = ModeRules(
     minimums={"nm": 4, "hd": 2, "hr": 2, "dt": 2, "tb": 1},
     optional_categories=("fm",),
     category_order=("nm", "hd", "hr", "dt", "fm", "tb"),
-    slot_mods={"nm": (), "hd": ("HD",), "hr": ("HR",), "dt": ("DT",), "fm": (), "tb": ()},
+    slot_mods={"nm": (), "hd": ("HD",), "hr": ("HR",), "dt": ("DT",), "fm": ("FreeMod",), "tb": ("FreeMod",)},
     allow_std_converts=False,
 )